requirements.txt
Django==1.9.7
django-filter==0.13.0
djangorestframework==3.3.3
drf-nested-routers==0.11.1
httpie==0.9.3
Markdown==2.6.6
Pillow==3.2.0
psycopg2==2.6.1
Pygments==2.1.3
requests==2.10.0
